Fix dependency tree (and update ESLint configuration for compatibility) (#5)
* Fix dependency tree (eslint version)

The eslint version in package.json was incompatible with
the ^7.11.0 requirement for the current react-scripts version.
Downgrading to ^7.32.0 solves this...

* Downgrade eslint-config-react

v19 was incompatible with react-scripts v4.0.3. Upgrading
react-scripts is beyond the scope of our PR.

* Update ESLint config for increased compatibility

Updates ESLint config for increased compatibility with
current practices in the codebase and to eliminate compilation
errors.

* Add newline at the end

*ahem*

* Add some ignore patterns to the eslint config

* Minimise changes to depencies per yarn.lock

* Fix dependency conflicts